requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.28.0
stravalib>=1.6.0
google-api-python-client>=2.128.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
pydantic-settings>=2.2.0
httpx>=0.27.0
python-dotenv>=1.0.0

# dev / test
pytest>=8.2.0
pytest-asyncio>=0.23.0
